refactor: add outer class name for `Avatar` and separate types
